\name{tm.lexicodeR-package}
\alias{tm.lexicodeR-package}
\alias{tm.lexicodeR}
\docType{package}
\title{
\packageTitle{tm.lexicodeR}
}
\description{
\packageDescription{tm.lexicodeR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tm.lexicodeR}
\packageIndices{tm.lexicodeR}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{tm.lexicodeR}

Maintainer: \packageMaintainer{tm.lexicodeR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
